+++ example/lib/screens/presentation_screen.dart
@@ -31,9 +31,10 @@ class _PresentationScreenState extends State<PresentationScreen> {
           controller: _controller,
           onRefresh: () => Future.delayed(const Duration(seconds: 2)),
           child: DecoratedBox(
-            decoration: const BoxDecoration(
-                color: appBackgroundColor,
-                boxShadow: [BoxShadow(color: Colors.black26, blurRadius: 8, spreadRadius: 4)]),
+            decoration:
+                const BoxDecoration(color: appBackgroundColor, boxShadow: [
+              BoxShadow(color: Colors.black26, blurRadius: 8, spreadRadius: 4)
+            ]),
             child: CustomScrollView(
               physics: const AlwaysScrollableScrollPhysics(
                 parent: ClampingScrollPhysics(),
@@ -56,7 +57,8 @@ class _PresentationScreenState extends State<PresentationScreen> {
                             width: double.infinity,
                             decoration: const BoxDecoration(
                               color: Colors.white,
-                              borderRadius: BorderRadius.all(Radius.circular(20)),
+                              borderRadius:
+                                  BorderRadius.all(Radius.circular(20)),
                               boxShadow: [
                                 BoxShadow(
                                   color: Colors.black12,
@@ -146,19 +148,25 @@ class _PresentationScreenState extends State<PresentationScreen> {
                                     const Spacer(),
                                     ElevatedButton(
                                       style: ButtonStyle(
-                                        backgroundColor: WidgetStateProperty.all(
-                                          _controller.isRefreshEnabled ? Colors.red : Colors.lightGreen,
+                                        backgroundColor:
+                                            WidgetStateProperty.all(
+                                          _controller.isRefreshEnabled
+                                              ? Colors.red
+                                              : Colors.lightGreen,
                                         ),
                                       ),
                                       child: Text(
-                                        _controller.isRefreshEnabled ? "DISABLE" : "ENABLE",
+                                        _controller.isRefreshEnabled
+                                            ? "DISABLE"
+                                            : "ENABLE",
                                         style: const TextStyle(
                                           color: Colors.white,
                                         ),
                                       ),
-                                      onPressed: () => _controller.isRefreshEnabled
-                                          ? _controller.disableRefresh()
-                                          : _controller.enableRefresh(),
+                                      onPressed: () =>
+                                          _controller.isRefreshEnabled
+                                              ? _controller.disableRefresh()
+                                              : _controller.enableRefresh(),
                                     ),
                                   ],
                                 ),
@@ -181,7 +189,9 @@ class _PresentationScreenState extends State<PresentationScreen> {
             return AnimatedBuilder(
               animation: _controller,
               builder: (context, _) => Stack(
-                alignment: _controller.side.isBottom ? AlignmentDirectional.bottomStart : AlignmentDirectional.topStart,
+                alignment: _controller.side.isBottom
+                    ? AlignmentDirectional.bottomStart
+                    : AlignmentDirectional.topStart,
                 children: <Widget>[
                   Container(
                     height: 100,
@@ -198,8 +208,9 @@ class _PresentationScreenState extends State<PresentationScreen> {
                     ),
                   ),
                   Container(
-                    margin:
-                        controller.side.isBottom ? const EdgeInsets.only(bottom: 100) : const EdgeInsets.only(top: 100),
+                    margin: controller.side.isBottom
+                        ? const EdgeInsets.only(bottom: 100)
+                        : const EdgeInsets.only(top: 100),
                     width: double.infinity,
                     height: 50,
                     color: Colors.greenAccent,
@@ -217,7 +228,8 @@ class _PresentationScreenState extends State<PresentationScreen> {
                   Transform.translate(
                     offset: Offset(
                       0.0,
-                      (_controller.side.isBottom ? -100 : 100) * _controller.value,
+                      (_controller.side.isBottom ? -100 : 100) *
+                          _controller.value,
                     ),
                     child: child,
                   ),